Code to probe the sensor value infrequently

+++ levitation.ino
@@ -5,11 +5,11 @@
 int sensor_pin = A0;  int sensor_value = 0;
 int error = 0;        int last_error = 0;  // Will be used for lead compensation
 int output = 0;       int last_output = 0;  // Will be used for lead compensation
-uint8_t test_output = 0;  // For testing loop cycle time
+uint16_t test_output = 0;  // For testing loop cycle time
 int pwm_value = 0;
 
 void setup() {
-//  Serial.begin(9600);  // Uncomment for debugging
+  //Serial.begin(9600);  // Uncomment for debugging
   //No prescaling on PWM clock
   TCCR0B |= _BV(CS00);
   TCCR0B &= ~(_BV(CS01) | _BV(CS02));
@@ -33,23 +33,14 @@ void loop() {
     // Lead compensation hump at 0.0002 sampling period
   //output =  (1) * error -   (2) * last_error +  (3) * last_output;
     //output = 9.671 * error - 9.598 * last_error + 0.9269 * last_output;  // hump = 120, K = 1
-        // Small continuous escillations, ~2mm. Occasionally these grow until it drops. Hangs higher than I'd expect
 //    output = 4.835 * error - 4.799 * last_error + 0.9269 * last_output;  // hump = 120, K = 0.5
-        // Large oscillations when partially supported by finger. Falls out of the sky on its own, no balancing
 //    output = 19.34 * error - 19.2 * last_error + 0.9269 * last_output;   // hump = 120, K = 2
-        // Really cool. You can clearly see the small fast oscillations build into slower larger ones before it crashes
 //    output = 37.46 * error - 36.9 * last_error + 0.8589 * last_output;   // hump = 240, K = 4
-        // Always unstable, medium sized oscillations and then falls
 //    output = 18.73 * error - 18.45 * last_error + 0.8589 * last_output;  // hump = 240, K = 2
-        // Stable! Like, really stable!
 //    output = 74.92 * error - 73.79 * last_error + 0.8589 * last_output;  // hump = 240, K = 8
-        // Fast, medium sized oscillations, drops as soon as support is removed
 //    output = 9.365 * error - 9.224 * last_error + 0.8589 * last_output;  // hump = 240, K = 1
-        // Stable! Like, really stable!
 //    output = 28.1 * error - 27.67 * last_error + 0.8589 * last_output;  // hump = 240, K = 3
-        // Vibrates constantly small and fast, but stays in place
 //    output = 14.05 * error - 13.84 * last_error + 0.8589 * last_output;  // hump = 240, K = 1.5
-        // Stable! Like, really stable!
 //    output = 2.341 * error - 2.306 * last_error + 0.8589 * last_output;  // hump = 240, K = 0.25
 //    output = 2.81 * error - 2.767 * last_error + 0.8589 * last_output;  // hump = 240, K = 0.3
 //    output = 3.278 * error - 3.228 * last_error + 0.8589 * last_output;  // hump = 240, K = 0.35
@@ -107,6 +98,11 @@ void loop() {
 
 //  Serial.print("Sensor voltage: "); Serial.println(sensor_value);  // Uncomment for debugging
 
+//  test_output++;
+//  if (test_output == 10000) {
+//    test_output = 0;
+//    Serial.print("Sensor voltage: "); Serial.println(sensor_value);  // Uncomment for debugging
+//  }
   last_error = error;
   last_output = output;
 }
